move out of utils.util plus cleanup
