+++ scilpy/io/utils.py
@@ -19,7 +19,7 @@ import six
 
 from scilpy.gradients.bvec_bval_tools import DEFAULT_B0_THRESHOLD
 from scilpy.utils.filenames import split_name_with_nii
-from scilpy.utils.util import RAS_AXES_NAMES
+from scilpy.utils.spatial import RAS_AXES_NAMES
 
 
 eddy_options = ["mb", "mb_offs", "slspec", "mporder", "s2v_lambda", "field",
