+++ scilpy/reconst/utils.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import warnings
-
 from dipy.direction.peaks import peak_directions
-from dipy.reconst.shm import sph_harm_lookup
 import numpy as np
